requirements.txt
pyyaml>=5.1
stomp.py
importlib-metadata